requirements.txt
# Core data processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0

# Data visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
altair>=5.0.0

# Machine Learning
scikit-learn>=1.3.0
statsmodels>=0.14.0
prophet>=1.1.0
xgboost>=1.7.0
lightgbm>=4.1.0

# Deep Learning
torch>=2.0.0
pytorch-forecasting>=0.10.0
tensorflow>=2.13.0

# Time Series
tsfresh>=0.20.0
sktime>=0.20.0

# MLOps
mlflow>=2.8.0
prefect>=2.14.0

# Generative AI
langchain>=0.0.300
langchain-community>=0.0.20
openai>=1.0.0
chromadb>=0.4.0

# API & Deployment
fastapi>=0.104.0
uvicorn>=0.24.0
streamlit>=1.28.0
docker>=6.1.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
jupyter>=1.0.0
notebook>=7.0.0
tqdm>=4.66.0